Modal shifted to new BtcInvoiceModal component
